feat: branching bonding period options
