+++ src/data/uiConfig.ts
@@ -1,5 +1,14 @@
 import { CellarNameKey, ConfigProps } from "./types"
 
+interface BondingPeriod {
+  title: string
+  amount: string
+  value: BondingValueOptions
+  checked?: boolean
+}
+
+type BondingValueOptions = 0 | 1 | 2
+
 export const isBondingEnabled = (config: ConfigProps) => {
   return true
 }
@@ -134,4 +143,73 @@ export const isUsdcChartEnabled = (config: ConfigProps) => {
     config.cellarNameKey === CellarNameKey.STEADY_BTC ||
     config.cellarNameKey === CellarNameKey.STEADY_ETH
   )
+}
+
+export const bondingPeriodOptions = (
+  config: ConfigProps
+): BondingPeriod[] => {
+  if (config.cellarNameKey === CellarNameKey.AAVE) {
+    return [
+      {
+        title: "7 Days",
+        amount: "1.1x SOMM",
+        value: 0,
+      },
+      {
+        title: "14 Days",
+        amount: "1.3x SOMM",
+        value: 1,
+      },
+      {
+        title: "21 Days",
+        amount: "1.5x SOMM",
+        value: 2,
+      },
+    ]
+  }
+  if (
+    config.cellarNameKey === CellarNameKey.ETH_BTC_MOM ||
+    config.cellarNameKey === CellarNameKey.ETH_BTC_TREND
+  ) {
+    return [
+      {
+        title: "10 Days",
+        amount: "1.1x SOMM",
+        value: 0,
+      },
+      {
+        title: "14 Days",
+        amount: "1.2x SOMM",
+        value: 1,
+      },
+      {
+        title: "20 Days",
+        amount: "1.25x SOMM",
+        value: 2,
+      },
+    ]
+  }
+  if (
+    config.cellarNameKey === CellarNameKey.STEADY_BTC ||
+    config.cellarNameKey === CellarNameKey.STEADY_ETH
+  ) {
+    return [
+      {
+        title: "10 Days",
+        amount: "1.1x SOMM",
+        value: 0,
+      },
+      {
+        title: "14 Days",
+        amount: "1.2x SOMM",
+        value: 1,
+      },
+      {
+        title: "20 Days",
+        amount: "1.25x SOMM",
+        value: 2,
+      },
+    ]
+  }
+  return []
 }
